fix: (platform) fix Display List Item examples for stackblitz
