+++ apps/docs/src/app/platform/component-docs/platform-display-list-item/platform-display-list-item-docs.component.html
@@ -10,8 +10,8 @@
     to bind the list items.
     Where we can pass list items individually where a number of items are less</description>
 <component-example>
-    <fdp-borderless-display-list-item-example>
-    </fdp-borderless-display-list-item-example>
+    <fdp-platform-display-list-item-border-less-example>
+    </fdp-platform-display-list-item-border-less-example>
 </component-example>
 <code-example [exampleFiles]="borderLessDLI"></code-example>
 <separator></separator>
@@ -24,8 +24,8 @@
     <code>[link]</code> will help to get into the desired destination on click of navigation. Below example shows
     declarative way to use it.</description>
 <component-example>
-    <fdp-display-list-item-with-navigation-example>
-    </fdp-display-list-item-with-navigation-example>
+    <fdp-platform-display-list-item-with-navigation-example>
+    </fdp-platform-display-list-item-with-navigation-example>
 </component-example>
 <code-example [exampleFiles]="dliWithNavigation"></code-example>
 <separator></separator>
