+++ apps/docs/src/app/platform/component-docs/platform-display-list-item/platform-display-list-item-docs.component.ts
@@ -3,10 +3,10 @@ import { Component } from '@angular/core';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
 
 import * as dliSrc from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-example.component.html';
-import * as borderLessDLISrc from '!raw-loader!./platform-display-list-item-examples/platform-borderless-display-list-item-example.component.html';
-import * as borderLessDLITs from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-border-less-example.component.ts';
+import * as borderLessDLISrc from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-border-less-example.component.html';
+import * as borderLessDLITs from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-border-less-example.component';
 import * as dliWithNavigation from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-with-navigation-example.component.html';
-
+import * as dliWithNavigationTs from '!raw-loader!./platform-display-list-item-examples/platform-display-list-item-with-navigation-example.component.ts';
 
 @Component({
     selector: 'app-standard-list-item',
@@ -25,13 +25,13 @@ export class PlatformDisplayListItemDocsComponent {
         {
             language: 'html',
             code: borderLessDLISrc,
-            fileName: 'platform-borderless-display-list-item-example',
+            fileName: 'platform-display-list-item-border-less-example',
         },
         {
             language: 'typescript',
-            component: 'PlatformDisplayListBorderLessExampleComponent',
+            component: 'PlatformDisplayListItemBorderLessExampleComponent',
             code: borderLessDLITs,
-            fileName: 'platform-borderless-display-list-item-example'
+            fileName: 'platform-display-list-item-border-less-example'
         }
     ];
 
@@ -40,6 +40,12 @@ export class PlatformDisplayListItemDocsComponent {
             language: 'html',
             code: dliWithNavigation,
             fileName: 'platform-display-list-item-with-navigation-example',
+        },
+        {
+            language: 'typescript',
+            component: 'PlatformDisplayListItemWithNavigationExampleComponent',
+            code: dliWithNavigationTs,
+            fileName: 'platform-display-list-item-with-navigation-example'
         }
     ];
